Feat: Refactor find inactive workplaces (#3100)
* feat(refactor-find-inactive-workplaces): Separate inactive workplace logic from next template logic

* feat(refactor-find-inactive-workplaces): Move nextEmail

* feat(refactor-find-inactive-workplaces): Re-structure non route files

* feat(refactor-find-inactive-workplaces): Remove need for variable
