+++ server/routes/admin/email-campaigns/inactive-workplaces/index.js
@@ -3,8 +3,8 @@ const router = express.Router();
 const moment = require('moment');
 
 const models = require('../../../../models');
-const findInactiveWorkplaces = require('../../../../models/email-campaigns/inactive-workplaces/findInactiveWorkplaces');
-const sendEmail = require('./sendEmail');
+const findInactiveWorkplaces = require('../../../../services/email-campaigns/inactive-workplaces/findInactiveWorkplaces');
+const sendEmail = require('../../../../services/email-campaigns/inactive-workplaces/sendEmail');
 
 const getInactiveWorkplaces = async (_, res) => {
   const inactiveWorkplaces = await findInactiveWorkplaces.findInactiveWorkplaces();
